pixvm: move A/B color channels to _Float16
Channels (the register file `pixv`) are now _Float16x8 instead of float32x8 --
native 128-bit NEON at 8 wide, half the register-file bytes, and the project's
existing color type (cnvs_premul/cnvs_unpremul). u8 <-> channel quantization
stays in float (the divide and *255+0.5 are exact there; fp16's [0,1] resolution
makes the u8 round-trip lossless), matching how the rest of the project converts
color. No geometric coordinates arise in this color/coverage ISA, so nothing
stays float32 yet. Shared pixvm_pixio.h does the conversions; SPLAT narrows the
float immediate; scalar-tail load/store widen with explicit casts (-Wdouble-
promotion). Tests still pass exactly (fp16 round-trip is lossless, source-over
values are exact in fp16).

Effect (256x256 source-over, -Os, A/A control 1.00x): fp16 sped the *unchecked*
work ~20-25% (A 57->43 ms, B 60->50 ms) but barely moved the bounds-safe build
(A 62->60, B 94->88), so the relative -fbounds-safety cost rose -- A 1.09x ->
1.40x, B 1.55x -> 1.78x. The integer index/pointer checks don't shrink with the
color data, so once the arithmetic is cheaper they're a bigger slice (same
mechanism as -O2 widening the gap). A-vs-B unchanged: threaded still ~1.5x slower
than switch under the flag.

+++ src/pixvm.h
@@ -5,9 +5,11 @@
 // ISA; pixvm_run_switch (one big switch) and pixvm_run_threaded (tail-call threaded)
 // are two backends, with an SkRasterPipeline-style register pipeline to follow.
 //
-// A run walks `n` pixels PIXVM_N at a time.  Channels are float in [0,1] held in a
-// register file of PIXVM_N-wide vectors; load/store ops convert to and from
-// tightly packed RGBA8.
+// A run walks `n` pixels PIXVM_N at a time.  Colour channels are _Float16 in [0,1]
+// (the project's colour type, and native 128-bit NEON at 8 wide), held in a
+// register file of PIXVM_N-wide vectors; load/store ops convert to and from tightly
+// packed RGBA8 with the u8 quantization done in float.  Any geometric coordinate
+// would stay float32 -- none arise in this colour/coverage ISA.
 
 #include <ptrcheck.h>
 #include <stdint.h>
@@ -15,7 +17,7 @@
 #define PIXVM_N 8       // pixels per step (vector lanes)
 #define PIXVM_REGS 16   // vector registers in the file
 
-typedef float pixv __attribute__((ext_vector_type(PIXVM_N)));
+typedef _Float16 pixv __attribute__((ext_vector_type(PIXVM_N)));
 
 // LOAD_SRC/LOAD_DST/STORE treat `dst` as the base of a four-register r,g,b,a group;
 // the arithmetic ops read `a`/`b`/`c` and write `dst`; SPLAT writes `dst` from imm.
